refactoring crud process job code
